migrate.sh: Check if migration even started and only kill process if it exists. start_isduba.sh: check if port 5173 is free before starting frontend.

+++ docs/scripts/migrate.sh
@@ -21,4 +21,9 @@ done
 
 mv isduba.log isduba_migrate.log
 
-kill $!
+if ps -p $!; then
+  echo "Migration successful"
+  kill $!
+else
+ echo "Couldn't start migration, is isdubad already running?"
+fi
